FloatWritable

+++ src/main/java/ca/uwaterloo/cs/bigdata2016w/xeniaqian94/assignment1/PairsPMI.java
@@ -234,9 +234,9 @@ public class PairsPMI extends Configured implements Tool {
 		}
 	}
 
-	protected static class SecondPartitioner extends Partitioner<PairOfStrings, IntWritable> {
+	protected static class SecondPartitioner extends Partitioner<PairOfStrings, FloatWritable> {
 		@Override
-		public int getPartition(PairOfStrings key, IntWritable value, int numReduceTasks) {
+		public int getPartition(PairOfStrings key, FloatWritable value, int numReduceTasks) {
 			return (key.getLeftElement().hashCode() & Integer.MAX_VALUE) % numReduceTasks;
 		}
 	}
